requirements.txt
# File: requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
beautifulsoup4>=4.9.0
feedparser>=6.0.0

# Machine Learning & NLP
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
nltk>=3.6.0
faiss-cpu>=1.7.0

# Financial APIs
yfinance>=0.2.0
alpha-vantage>=2.3.0

# Document Processing  
PyPDF2>=3.0.0
python-docx>=0.8.11
mammoth>=1.4.0

# Web Frameworks (for future API/UI)
streamlit>=1.25.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Utilities
python-dotenv>=0.19.0
schedule>=1.1.0
tqdm>=4.62.0
python-dateutil>=2.8.0

# Optional enhanced features
umap-learn>=0.5.0  # For dimensionality reduction
plotly>=5.0.0      # For visualization
dash>=2.0.0  